undo changes to other files
